requirements.txt
# Core Dependencies
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.3.0

# Data Collection
requests>=2.31.0
aiohttp>=3.8.0
yfinance>=0.2.28
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Database
sqlalchemy>=2.0.0
sqlite3-python>=1.0.0

# Visualization
plotly>=5.15.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Web Application
streamlit>=1.25.0
dash>=2.11.0
dash-bootstrap-components>=1.4.0

# NLP & Sentiment Analysis
vaderSentiment>=3.3.2
textblob>=0.17.1
nltk>=3.8.0
spacy>=3.6.0
transformers>=4.30.0

# AI Integration
google-generativeai>=0.3.0
openai>=1.0.0
langchain>=0.1.0

# Portfolio Optimization
cvxpy>=1.3.0
pyportfolioopt>=1.5.0
quantlib>=1.30

# Backtesting
backtrader>=1.9.78.123
zipline-reloaded>=2.4.0
vectorbt>=0.25.0

# Data Processing
python-dateutil>=2.8.2
pytz>=2023.3
holidays>=0.30

# Configuration
python-dotenv>=1.0.0
pyyaml>=6.0
configparser>=6.0.0

# Utilities
tqdm>=4.65.0
joblib>=1.3.0
click>=8.1.0
colorama>=0.4.6

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
mock>=5.1.0

# Development
black>=23.0.0
flake8>=6.0.0
mypy>=1.4.0
ipython>=8.14.0
jupyter>=1.0.0

# Logging & Monitoring
loguru>=0.7.0
sentry-sdk>=1.28.0

# Performance
numba>=0.57.0
cython>=3.0.0

# API Rate Limiting
ratelimit>=2.2.1
tenacity>=8.2.0

# Crypto-specific
ccxt>=4.0.0
web3>=6.0.0

# News API
newsapi-python>=0.2.6
feedparser>=6.0.10

# Technical Analysis
ta>=0.10.2
ta-lib>=0.4.27

# Time Series
statsmodels>=0.14.0
prophet>=1.1.0
arch>=6.1.0

# Optional GPU Support (uncomment if needed)
# tensorflow>=2.13.0
# torch>=2.0.0
# torchvision>=0.15.0